\name{asSeq2-package}
\alias{asSeq2-package}
\alias{asSeq2}
\docType{package}
\title{
\packageTitle{asSeq2}
}
\description{
\packageDescription{asSeq2}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{asSeq2}
\packageIndices{asSeq2}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{asSeq2}

Maintainer: \packageMaintainer{asSeq2}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
